requirements.txt
# Base de données
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# Utilitaires
python-dateutil==2.8.2